requirements.txt
python-telegram-bot==13.7
numpy==1.21.4
pandas==1.3.3
openai==0.28